local_docker/health-nutrition-api/requirements.txt
flask
phidata
pandas
pyyaml
# pip install groq langflow